Changed the computational scheme of the effective matsubara coupling matrix, which should now be much more efficient. Also created a small access check for Matzubara splitting vector. Working on the Terminal output. Error and Warning should derive from terminal out to not have dublicate code.

+++ scallop/eliashberg/MatzubaraSplittingVector.h
@@ -30,9 +30,9 @@ template<typename T>
 class MatzubaraSplittingVector : public MatzubaraSplittingBase<MatzubaraSplittingVector<T>,T> {
 public:
 
-	T & operator() (size_t b, size_t j, size_t n);
+	T & operator() (size_t b, size_t j, int n);
 
-	T operator() (size_t b, size_t j, size_t n) const;
+	T operator() (size_t b, size_t j, int n) const;
 
 };
